+++ src/arm/linux/init.c
@@ -335,7 +335,7 @@ void cpuinfo_arm_linux_init(void) {
 #endif
 
 #if CPUINFO_ARCH_ARM
-	uint32_t isa_features = 0;
+	uint64_t isa_features = 0;
 	uint64_t isa_features2 = 0;
 #ifdef __ANDROID__
 	/*
@@ -388,7 +388,7 @@ void cpuinfo_arm_linux_init(void) {
 		&chipset,
 		&cpuinfo_isa);
 #elif CPUINFO_ARCH_ARM64
-	uint32_t isa_features = 0;
+	uint64_t isa_features = 0;
 	uint64_t isa_features2 = 0;
 	/* getauxval is always available on ARM64 Android */
 	cpuinfo_arm_linux_hwcap_from_getauxval(&isa_features, &isa_features2);
